feat: TET-863 move stylesBuilder into main component folder
